Actually dereference views in LayoutGuideCenter
Prior to this CL, views were not dereferenced. The underlying storage
for views is NSMapTable and its setObject(forKey:) doesn’t remove the
key when the value is nil (contrary to Dictionary which was used in a
prior implementation).

This CL removes the key when the view argument is nil.

Bug: 1366724
Change-Id: I574c10e8b82301ee7a29457c5187bb5c8c7b4f96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3909589
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1050152}
NOKEYCHECK=True
GitOrigin-RevId: ce1033f9f55d16e0459a715e2412a815b259bc37