Reland "[iOS] Enable New Overflow Menu followup experiments on ToT"
This is a reland of commit 303933c5c43d74e4a7fc7fe6562032bfd2795bb7

The original CL was reverted because the tests failed on iOS 14.
This is because the overflow menu (and thus the new item) never work
on iOS 14, so the count was wrong.

Original change's description:
> [iOS] Enable New Overflow Menu followup experiments on ToT
>
> Launching these experiments has been approved. We decided not to launch
> the New Settings Action, so that one is not enabled.
>
> Bug: 1231546
> Change-Id: Icc1b7a7c300cc795130a335b5ddf05303cdb2a23
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3909749
> Reviewed-by: Guillaume Jenkins <gujen@google.com>
> Commit-Queue: Robbie Gibson <rkgibson@google.com>
> Cr-Commit-Position: refs/heads/main@{#1050266}

Bug: 1231546
Change-Id: Iafb372850aa87bb4beb3b39492dd64f13abf046e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914309
Auto-Submit: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/main@{#1050678}
NOKEYCHECK=True
GitOrigin-RevId: 10e4f96b735767c3af985c6cb2e5f285a2cf0e5b