[ios] Deregister the post device restore sign-in promo if not needed
The PromosManager stores registered promos in local prefs, which means
registered promos will survive an app restart. But for this promo, we
want to evaluate the conditions at startup each time, to determine if
we should display the promo. That means that we need to deregister
them if they should not be displayed.

Bug: 1362029
Change-Id: If9585c19b2922cb7fb77ea5439fb81e20f693df9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3905967
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1051176}
NOKEYCHECK=True
GitOrigin-RevId: 3d8c48ade6ba36b39d30b1d17e6712336089f68a