[iOS] Add new cache size for avatar
This patch renames IdentityAvatarSize::DefaultLarge to
IdentityAvatarSize::Regular.
And adds IdentityAvatarSize::Large for 48px size.
This size will be used for TangibleSync.

The avatar with its border is 60x60 in the mocks, but the avatar itself
is 48px for tangible sync.

  Mocks:
https://docs.google.com/presentation/d/10ceBtNVnUn3YKA02Z8NJmbT7wV9xKNKFoc3YghZGFog/edit?resourcekey=0-8QmaD0ZuVro4Ohpm43ooaw#slide=id.g13aee3cb85f_1_100
  Design dock:
http://go/ios-tangible-sync-design-doc

Bug: 1363812
Change-Id: I0362cd12fa00af4ac6f52e1d86f1b04c136c8704
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913348
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051239}
NOKEYCHECK=True
GitOrigin-RevId: be50e83c57828df61a19463393915f7f1aece2bf