Add a random suffix to background download session identifier
Although background download only creates one URL session. But the session is not invalidated synchronously internally in the OS. As a result, unit tests see a second session is created with the same identifier.
This CL fixes the issue by adding a random suffix  to the identifier. So 2 tests will not have the same identifier.

BUG=1367328

Change-Id: I3a632a61c5c5bdc225fd2dfe4f8c54c42ec73ee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3916598
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Min Qin <qinmin@chromium.org>
Commit-Queue: Min Qin <qinmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051325}
NOKEYCHECK=True
GitOrigin-RevId: 57b1adf44424749678ead41cf6739e00fd471715