[ios] Properly shutdown FeedTopSectionMediator
FeedTopSectionMediator registers an IdentityManager observer
which needs to be disconnected when the mediator is no longer
in use by calling the -shutdown method.

Previously this was left to be called when the object is deallocated
which can be arbitrarily delayed by animation and the autorelease
pool. Add an explicit call to -shutdown when the coordinator is
stopped (as it is the one that creates the mediator) to ensure the
unregistration happens at a known point in time.

Fixes a shutdown crash where the observer was not unregistered
by the time the IdentityManager was destroyed.

Fixed: 1364209
Change-Id: I83c7a8fcb265e4c7edfb662199a82f303986de3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3921755
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051737}
NOKEYCHECK=True
GitOrigin-RevId: 21f38d6743332ee91fce69eacd6f31686dcc155c