[ios] Trigger the sign-in flow from the post restore sign-in promo
This implements the primary action handler for the post restore sign-in
promo. This primary action kicks off the sign-in and sync flow with the operation set to "Reauthenticate".

It also adds the IdentityButtonControl to display the user's pre-restore account details. This control is currently loaded with the account email address, but the intention is to eventually load the name and avatar as well.

Bug: 1363283
Change-Id: Ieb72b86599588aa6319f1e8012ef159c1a8aa532
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3897339
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1051768}
NOKEYCHECK=True
GitOrigin-RevId: 090c1b2b8abf9e5b0131c555edf9663ae0810ad8