Remove postCrashLaunch check from PromosManagerSceneAgent.
Currently, PromosManagerSceneAgent checks against appState's
postCrashLaunch to determine if a promo should be displayed or not.

Unfortunately, iOS kills applications very frequently, and checking
against postCrashLaunch to determine if the app is launching after a
crash is not reliable.

This change removes this if-else check.

Bug: 1370075
Change-Id: Ib3f624e111ad468c1a1bc57c673bb4a52cf839ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3930245
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Cr-Commit-Position: refs/heads/main@{#1053709}
NOKEYCHECK=True
GitOrigin-RevId: 98e7aa8580561553af739b3ca41e66252356eff0