[PWM UI Split] Re-add On-device Encryption items
This CL adds the on-device encryption row (in various states) to the new
Password Settings submenu. This is the last new item for this UI.

As before, this CL does not include mediator changes; placeholder data
is provided to the consumer. I have tested this CL by (a) dispatching
various calls to setOnDeviceEncryptionState on a delay and ensuring the
UI updates correctly, and (b) working on the follow-up CL in parallel.

The follow-up CL will include unit and EG tests adapted from the parent
directory, once the mediator is ready.

LOW_COVERAGE_REASON=coming with follow-up mediator CL

Bug: 1335156
Change-Id: I32ece56a726d3f4156fcc38910441857cfceb044
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3926255
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Ernesto Izquierdo Clua <eic@google.com>
Cr-Commit-Position: refs/heads/main@{#1053718}
NOKEYCHECK=True
GitOrigin-RevId: f408943db64855a9ad21e94d1037478a540d0bb1