[iOS] Bug fix: "Passwords in other apps" text appear out of the cell
Issue: With dynamic large text, if the user goes to Settings->Password
Manager, the text "Passwords in Other Apps" appears out of bounds of the
cell: https://screenshot.googleplex.com/ANqQEoiRuHkc7c3.png

Root cause: The text "on/off" is not added when the cell first loads,
causing the cell to have a smaller instrinct height. When the detailed
text is retrieved and displayed, the width of the "passwords in other
apps" label is shrank as a result, and caused its height to increase.
However, the cell is not redrawn to accommodate the new height.

https://screenshot.googleplex.com/6uFvbUpii9dAUtG.png

Fix: Use `reloadCellsForItems` instead of `reconfigureCellsForItems` to
redraw the cell when detailed text is retrieved.

Fixed: 1369595
Change-Id: I46ffa4353212cb8d8b6bc479da9b71c544582c38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3925962
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1054704}
NOKEYCHECK=True
GitOrigin-RevId: 1d55c3c9711207941e8416297b2e0f483bb33a86