Create feature flag for App Store Rating promo testing.
This change creates a feature flag that will be used to manually test
the App Store Rating promo. The flag will be used to set the conditions
that will make the user eligible to be shown the promo (to be implemented
in a future CL).

Bug: 1359065
Change-Id: I3c02594b3849616cfbfb60a47726f2b4fc262990
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3936502
Commit-Queue: Hira Mahmood <hiramahmood@google.com>
Reviewed-by: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1055471}
NOKEYCHECK=True
GitOrigin-RevId: 341ddd6230fe98f3277bc23502f3b6d6c0c2cfcb