Add feature flag for Keyboard Shortcuts Menu
This CL adds a feature flag for the Keyboard Shortcuts Menu. The menu is built
when the AppDelegate is a UIResponder. This CL makes MainApplicationDelegate be
a UIResponder. In the framework call (-buildMenuWithBuilder:), the feature flag
is checked and conditionally builds the menu.

Context: go/bling-shortcuts-menu
Bug: 1371848
Change-Id: I3da8f5fc6f1290787b16eba98f606cc8c107295e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3934848
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1056248}
NOKEYCHECK=True
GitOrigin-RevId: c18612b96e9224451a998fca2e4adc648e3d73c9