[ios][finch-first-run] Adds skeleton for variations seed app agent
This CL adds the new variations seed fetching init stage, and the
skeleton of the app agent that controls it. As of now, the dynamic fre
finch feature is hardcoded to NO (the setup will be implemented in a
follow up CL) so it will always move to the next init stage immediately,
but the skeleton outlines a general workflow how it would handle the
app init stage changes and scene state changes.

Change-Id: I0512c1ade2d2496641a74310491f782e3a668823
Bug: 1372180
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3938264
Reviewed-by: Mark Cogan <marq@chromium.org>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057133}
NOKEYCHECK=True
GitOrigin-RevId: 3e6f25203402182e617531929264f70f24a22f51