Implement Feature-level Push Notification Toggling
In this CL, the prefService was extended to include a pref dictionary
that maps a PushNotificationClient to a boolean that indicates whether
the users wants to receive notifications from this feature. In addition,
PushNotificationService's UpdateFeaturePushNotificationPreferences
method signature was defined.

This is a multi-part CL:
  -  cl/478562375
  => https://crrev.com/c/3933549
  -  https://crrev.com/i/5013119

Bug: 1369616
Change-Id: I3504468111913ad1e17bf997c4e116ec29f97e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3933549
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057627}
NOKEYCHECK=True
GitOrigin-RevId: 53d3525c8beb761980614faad7271a846f43fbd2