[iOS] De-flake CRWWKHTTPCookieStoreTest.GetCookiesCachedCorrectly
GetCookiesCachedCorrectly is flaky because by calling SetCookies, it
sets up a race involving asynchronous messaging between the browser
process and the network process. At some point after setting cookies,
the network process sends a message which triggers the
|cookiesDidChangeInCookieStore:| callback, clearing the cache.
Typically, this happends before the first call to |getAllCookies:|
returns, so the second call to |getAllCookies:| is able to use the
cached value, as expected. But sometimes the callback occurs after
the first call to |getAllCookies:| returns. In this case, the
cache gets invalidated and isn't used for the second call, causing
the test to fail.

This CL fixes this flake by removing the call to SetCookies.

Bug: 1272949
Change-Id: I75fd495b0c5d6cf3b755eb33769f0879bf3715f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3934408
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Ali Juma <ajuma@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1057633}
NOKEYCHECK=True
GitOrigin-RevId: 2c33926873410184c18a47b101f2ca719a61f6d0