Fix NTPTileSaverControllerTest on iOS 16.1+
iOS 16.1 has a different behaviour with UIImage-s created from
UIGraphics* APIs, where an in-memory image won't have the same data
representation as if it were loaded from disk.

This CL adds a step so that the mock image also makes a round-trip to
the disk.

LOW_COVERAGE_REASON=This adds a codepath that runs only on iOS 16.1+. The Coverage bot runs on something lower.

Bug: 1370580
Change-Id: I1badee133c73ea06a8cce569bbd734b28d6b6d9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3948726
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1058089}
NOKEYCHECK=True
GitOrigin-RevId: 0a53ff685f914290ab165dc22811652dacec651a