[ios] Store pre-restore account info in local prefs
If we detect that the app is starting after an iOS device restore and
we have lost the account token, instead of just storing the account info
in a global variable, this change will store the account info in prefs
so that we can use this info beyond just the first app session. One
reason to do this is that the first session after a device restore could
be the OS starting the app in the background and there may not be a
chance to display the post restore signin promo. By storing this pre-
restore account info in prefs, we can still display the post restore
signin promo on the first foreground session.

Bug: 1362029
Change-Id: I9db4809dc10debdb8787300c1c3521933850ab2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3933075
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1058209}
NOKEYCHECK=True
GitOrigin-RevId: 0da607a3ac32d414c6cac93b40f5481189c85af7