Create experimental flag to force App Store Rating Promo.
This change creates an experimental flag to override the Promos Manager
and App Store Rating trigger criteria to force the App Store Rating
promo to display when the app is launched or resumed.

In a later CL, the flag will be updated to support forcing
the Post Restore SignIn promos.

Bug: 1359065
Change-Id: I914a4ab8dedc47417f1ac2d9bbb6b669938090d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3939945
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Benjamin Williams <bwwilliams@google.com>
Commit-Queue: Hira Mahmood <hiramahmood@google.com>
Cr-Commit-Position: refs/heads/main@{#1058332}
NOKEYCHECK=True
GitOrigin-RevId: b1058ec205ef9ad3c04438ab02fcd49d4ca71e26