[iOS] Change the incognito symbol to be configured with "Palette"
This CL changes the incognito symbol in a circle to a new one which
colors can be configured through a Palette symbol configuration.
This allows to tint the background and the foreground of the image
separately directly from the code.

However this code is only available in iOS 15+, so this CL is also
adding pre-tinted assets for the badge the incognito view.

Bug: 1315544
Change-Id: Ibb3abc51c12092eea5c03454fbf55b45f910ff87
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3945084
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1058655}
NOKEYCHECK=True
GitOrigin-RevId: 157089f08dfedfa0686d6556063a0535ea717fbe