[iOS] Create folder before start download
When you try to download a file with a given path and this path does
not exist, it throw an error. This CL fix this by checking if the folder
exist or create it otherwise when user tap on the share button, before
starting the download.

Fixed: 1371047, 1373233, 1373230
Change-Id: Ibee5797f290b82e78d8ba72d3aa4ba977d36a685
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3948509
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Aliona Dangla <alionadangla@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1059189}
NOKEYCHECK=True
GitOrigin-RevId: 0ee2fb6db0754b2579419265bb64dbc4463410c3