[iOS] Log incognito reauth histogram after UI enabled
A new init stage that may delay the initialization of browser objects
(including self.localState) is being added. This caused a scenario that
nothing is ready yet when foreground becomes inactive (or even active)
and the DCHECK when logging the histogram will fail.

As the only constraints on logging this metric are "can log a metric"
and "only log once per launch", moving this method to
`sceneDidEnableUI` so that when it is logged, we can be sure that
`self.localState` is ready.

Bug: 1372180
Change-Id: I9f42f8d754798e2d76139dbd1ff2ac40e0f84922
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3960950
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1060477}
NOKEYCHECK=True
GitOrigin-RevId: 3673ec97ae2c0d3c391084f73f988a8f076e5770