Implemented Price Notifications Mediator Skeleton
In this CL, the boilerplate code for the Price Notificaitons mediator
and consumer APIs were created. In addition, the
PriceNotificationCoordinator was adapted to instantiate and store the
new mediator.

In a future CL, the mediator will be extended to display the items that
are trackable on the current page.

Bug: 1373071
Change-Id: Ib7103c8c61993afa09a5b85517e56a8546139ea3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3946688
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Daniel White <danieltwhite@google.com>
Cr-Commit-Position: refs/heads/main@{#1060536}
NOKEYCHECK=True
GitOrigin-RevId: d698e284ed0c5afc2419e9167cb870ddc1a8e7b4