Migrate kGoogleServicesLastAccountId to kGoogleServicesLastGaiaId
The old preference was most of the time a GaiaID, but could sometimes
(rarely) be an email.
This was in particular the case on old Ash profiles.

This is confusing and causing issues for the Lacros migration.

This CL fixes the pref to always be a GaiaID. If it was an email, it is
simply discarded.
Discarding the pref is not a problem:
- it is only used in to detect cross account Sync issues.
- only Ash had an email there, and Ash cannot change the sync account,
  so there cannot be cross account sync issues on that platform.
- Lacros may also have an email there (copied from Ash) on the very
  first run. There cannot be a cross sync account issue on first run.

Bug: 1340791
Change-Id: Iafc64bbdac2690bce9aa577395cb6effa696c5ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3944995
Auto-Submit: David Roger <droger@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1060924}
NOKEYCHECK=True
GitOrigin-RevId: 877bdc767b99a23cb6696d9a44a7831b0263722c