Migrate KeyCommandsProvider from block-based API
We rely on target actions again (as this is the only way for key
commands to be added to the app menu).

Important: the selectors for the actions are all prefixed with keyCommand_
to ensure they don't trigger unintentionally a responder earlier in the
chain. For example, BVC has a startVoiceSearch method [1]  that would
have triggered before KeyCommandsProvider.

[1]: https://source.chromium.org/chromium/chromium/src/+/main:ios/chrome/browser/ui/browser_view/browser_view_controller.h;l=166;drc=814d7fcab7a7b0304836951cbd61b2a4799bf4ae

Context: go/bling-shortcuts-menu

Bug: 1371848
Change-Id: Ib6859ae54a1b17810fff1f62068ce82aceb135cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3963151
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1062037}
NOKEYCHECK=True
GitOrigin-RevId: dad5b9b1db92c16b8596c903ce5ce34cfd4e146b