[iOS][RecentTabs] Resets 'searchTerms' after opening a tab
This CL sets 'searchTerms' to an empty string after opening a tab.

When opening a tab from the RecentTabs view, the tabGrid mode is set to
'TabGridModeNormal' and the searchBar is set to an empty string.
Therefore, it is also necessary to set 'searchTerms' to an empty string.

Fixed: 1369073
Change-Id: I6ba9853f04be02cce6bf87a380423ea14d4bebbf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3970919
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1062699}
NOKEYCHECK=True
GitOrigin-RevId: 996ca6ab6bceb02482c7c9bc655bb536d8921d5a