[iOS] Reenable testNewFolderNameCommittedWhenKeyboardDismissedOnIpad
The tests are passing on device. On simulator, it is not passing
because tapping on the "hide keyboard" key is not considered as
dismissing the keyboard (probably because the system consider the
hardware keyboard as still connected).

Fixed: 801453
Change-Id: Idd6f8cf70a0f76ada17e3bcb6510473201522293
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3970887
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Louis Romero <lpromero@google.com>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1062712}
NOKEYCHECK=True
GitOrigin-RevId: 2454b28ce5bcf9a1bd9bea2a1d0728896a13d5cf