[ios] Update link and payload used for Share Chrome feature
The new URL is a Google-owned URL (rather than a direct Apple app-store
link) and it redirects to the appropriate app store for the type of
device (ie, either the iOS app store or the Android Play store).

The code that builds an `LPLinkMetadata` was required in order to get
the UIActivityViewController to properly display a preview for direct app-store link. This is not necessary for the new Google-owned URL - a preview / thumbnail is properly generated automatically, so that code is no longer necessary.

The text payload is also updated to reflect that this new URL works
for more than just iOS.

Bug: 1376575
Change-Id: Icbd3b451fa19b7830d30f2c7c8e60e3d5559be45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3967040
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Scott Yoder <scottyoder@google.com>
Cr-Commit-Position: refs/heads/main@{#1062828}
NOKEYCHECK=True
GitOrigin-RevId: 20b374985f0a1fba298dbdb6318777a2cf434c0b