History: More reliably populate visit_duration
A visit's duration should be computed and stored once the visit is
"finished", either because the user navigated away or because they
closed the tab.
Before this CL, the "navigated away" case was handled directly in
HistoryBackend: When a new visit was added, the referrer visit's
duration got updated.
That failed to handle any cases where there is no referrer - e.g. due
to the site's referrer policy, or because the user typed a new URL
into the Omnibox.

This CL fixes the issue by moving the visit-duration computation into
HistoryTabHelper, which anyway had to detect new navigations (and
which already handled the "tab closed" case).

Bug: 1357013, b/255223185
Change-Id: I153eded809c2dc2ab9239f9552563c664d834e90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3856273
Reviewed-by: Sophie Chang <sophiechang@chromium.org>
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1063843}
NOKEYCHECK=True
GitOrigin-RevId: 28e4521445a88b1e18c98836f74b2e148f797b87