[ios] Convert ChromeEarlGrey to use base::TimeDelta
Using base::TimeDelta allow to specify the time unit when defining
the constant without resorting to putting the unit in the variable
name.

This also helps converting between different units with the helper
methods (.InSecondsF(), .InNanosecondsF(), ...).

This converts ChromeEarlGrey to use base::TimeDelta instead of
NSTimeInterval for the duration and convert all the client code.

This is part of a multi-CL change:
 -  https://crrev.com/c/3975698
 -  https://crrev.com/i/5054460
 => https://crrev.com/c/3971506
 -  https://crrev.com/i/5054461
 -  https://crrev.com/c/3975841

Bug: none
Change-Id: Ic6b4b43e7ad2fdaf0a9c2a391f70bf74dd20cfc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3971506
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1064299}
NOKEYCHECK=True
GitOrigin-RevId: 745dd20cf7d3648cc7d4599c1513a7a6e8da06d3