[iOS][finch-first-run] Split iOS variations manager to fetcher and store
Previously, to retrieve the downloaded first run seed, the variations
service client needs to initialize another instance of
IOSChromeVariationsSeedManager just to "pop" the seed. This CL refactors
the code so that the fetcher and the storage are in different classes,
and to retrieve the fetched seed, the variations service client only
needs to invoke a class method on the storage to do so.

Original comment: crrev/c/3852302/comment/44f3142c_01aa65c5/. At that
time I was suggested that I should use a free function, but there was
no previous cases where a free function definition and an ObjC
@interface declaration reside in the same header file, which feels
weird to me. Having a separate "store" class with setter methods only
exposed to the fetcher tallies more with my original intention :)

Also updated timeout time to 2s to comply with UX requirements.

Bug: 1353937
Change-Id: Idf9f8016818689845cc31ac1ca9445638c546e26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3970671
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1064381}
NOKEYCHECK=True
GitOrigin-RevId: d1fd8acd7721a8f243824030e8cf9d6cd6d2662d