[iOS] Exit fullscreen mode when permission prompt shows
These changes are behind a feature flag. When the feature flag is
enabled, if the user is in fullscreen mode, the user will exit
fullscreen mode before the permission prompt is shown. This needs to be
done since the user can't see the permission prompt or the following
infobar message in fullscreen mode.

Fixed: 1315263
Change-Id: Ifcf7006c1fb675915daaee10399249d0e94d0568
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3984665
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1064440}
NOKEYCHECK=True
GitOrigin-RevId: 14614bc731bad6499faec027e653f5b0461b7aec