[iOS] Remove failing DCHECK in didCommitNavigation
This removes a DCHECK that the navigation has a context when a
navigation is committed. Based on metrics and crash reports,
this DCHECK is known to fail in the wild, and there is already
code to handle the case where the DCHECK doesn't hold. Removing
logic to handle this case would produce a top-crash. The
Chromium style guide disallows having a DCHECK followed by logic
to handle the case where the DCHECK doesn't hold.

Bug: 1377750
Change-Id: I1d832dc633770cdd8d0b99c430cd58878c319b41
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3989003
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1064547}
NOKEYCHECK=True
GitOrigin-RevId: bf298c18a0f5335ec8fc134fff6cba42520cbbed