[iOS] Build menu from main app delegate
This CL register key commands only once on launch, that appear in the
app menu.

This need to be done at launch because at the runtime the OS ask the
responder chain to see if one of the registered key commands can be
performed to be displayed in the menu.

More explanation: go/bling-shortcuts-menu

See screens:
https://drive.google.com/file/d/1levc6bwz4ubBssA2mTAcLAjC29ezBqHQ/view?usp=sharing
https://drive.google.com/file/d/1V9Inxem7GUadJZButN-34hrzNUFhIdgh/view?usp=sharing

Fixed: 1376445
Change-Id: I82472867f64e48b5e6fd3f95111b3b3567c2a868
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3976338
Commit-Queue: Louis Romero <lpromero@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Cr-Commit-Position: refs/heads/main@{#1064551}
NOKEYCHECK=True
GitOrigin-RevId: e293fd192c5b5918b99cdc4befb151fb4999a56a