[history] Refactor history::ContextID to be a pointer-sized number
Goal: Stop the dangling pointer crash reports for History.

history::ContextID has always been a bit strange. It's basically a
pointer to a WebContentsUserData instance that's not supposed to be
dereferenced, but instead used as a unique identifier.

Chrome is adding automated tooling to detect dangling pointers, and
usage of history::ContextID is tripping that detector, even though it's
never actually dereferenced, and forbidden from being dereferenced by
a comment warning.

This CL makes history::ContextID an std::uintptr_t, which is a number
that's sized the same as a pointer, which makes it explicit that it
will never be dereferenced.

It also updates some callsites and removes some annotations that are
sending us crash reports.

Bug: 1376915, 1291138
Change-Id: Ia4bbdf958d78c5b9ed0fb4472eff6b099af5ddd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3988089
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Commit-Queue: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1064932}
NOKEYCHECK=True
GitOrigin-RevId: ed7ec4f7e1db62b94d6398fc7e521b42960e00fd