[iOS][finch-first-run] Expiry mechanism for variations app state agent
This CL adds the check whether there is a current seed available in the
variations seed store and whether it has expired. The result  will (in
later CL) be recorded by metrics. This mechanism is also used to
determine whether the app is in its first run.

Since the variations service and seed store objects would not be
available when this check executes, the seed fetch time is retrieved
from NSUserDefaults, instead of from the seed store. Every time the app
backgrounds, however, the app would retrieve the fetch time of the last
seed from variations seed store and puts it into NSUserDefault.

The above logic would be executed every time the app runs and shuts down
instead of only on the first run, as we previously designed it to be.
This is because the existing first-run-checker is unavailable when this
code is executed, and we have to resort to adding a new field in
NSUserDefaults anyways, so why not make it more useful?

For this purpose, new APIs are added to variations_seed_store.h. Since
this function imports `components/variations/", the DEP file is updated
to include it.

Bug: 1372180
Change-Id: Ice468664c17720cbe67ba40725f946b784b8b6b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3968540
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1066173}
NOKEYCHECK=True
GitOrigin-RevId: 5db2b1d55c31ef5b55d95da251492d3cd78a75b5