Add missing deps to some //ios factories using a script
This covers simple cases where the factory was directly used in the
file but not listed in DependsOn(). This doesn't cover cases where the
depended factory is used in a helper defined outside of the file.

Bug: 1379097
Change-Id: I8fe2682fe796a486658745d65f13f0a3915aafaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3996835
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1066424}
NOKEYCHECK=True
GitOrigin-RevId: 6bc8885e808941e1bbbe108b1da487a258b37317