[SafeBrowsing] Expose duration constant as base::TimeDelta
According to discussions on [cxx], it is okay to expose constexpr
constants as `inline constexpr` from components. Instead of having
a constant whose time unit is expressed in its name, change it to
a `base::TimeDelta` initialised with the constructor that specify
the correct unit.

Simplify TailoredSecurityTabHelper::OnTailoredSecurityBitChanged()
on iOS to avoid unnecessary computation when `enabled` is false or
the WebState is hidden.

Use base::Time::Now() instead of base::Time::NowFromSystemTime()
in the iOS unit tests since they don't depend on specific real-world
timing (as this is used to construct fake time point).

[cxx]: https://groups.google.com/a/chromium.org/g/cxx/c/hmyGFD80ocE/m/O4AXC93vAQAJ

Bug: none
Change-Id: I68f099e55bd041a135d32ee54bd863e7d2fbd1ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3986025
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1066461}
NOKEYCHECK=True
GitOrigin-RevId: 52457f117e8a0fa9797947eca1ac4701017f5042