[iOS] Fix official build failure
This CL reverts crrev/c/3993535 and fixes the root cause of the offcial
build failure
https://ci.chromium.org/ui/p/chrome/builders/official/ios/8147/overview.

The cause of the failure is that on official builds, fetching is enabled
by default. This is now overridden in the testing double.

Fixed: 1380141
Change-Id: I5182e89eeddf2b411fef2ffcd077565e8f966d7c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3999259
Reviewed-by: Daniel White <danieltwhite@google.com>
Commit-Queue: Daniel White <danieltwhite@google.com>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1066726}
NOKEYCHECK=True
GitOrigin-RevId: 78f30800199fa15d4f6a9f45a716d18fa84e6d3d