[ios] Remove -waitForWebStateContainingText:timeoutInSeconds: method
This was added to prevent breaking the internal repository when the
type of timeout in -waitForWebStateContainingText:timeout: changed.
As the internal repository has been fixed to pass a base::TimeDelta
it is now possible to remove this unused method.

This is part of a multi-CL change:
 -  https://crrev.com/c/3975698
 -  https://crrev.com/i/5054460
 -  https://crrev.com/c/3971506
 -  https://crrev.com/i/5054461
 => https://crrev.com/c/3975841

Bug: none
Change-Id: I999ebbc6fd256ca18bbd80eb7ef8766adebe5303
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3975841
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1067030}
NOKEYCHECK=True
GitOrigin-RevId: 0ee3cc418af2d68c15a5641a35b74727155d0407