[ios] Modified scroll condition for Good Visits UMA metric.
Changed the scroll condition for the metric. Before, the metric took would be triggered when the user scrolled after 60 seconds. Now it aligns with the original specification of at least 1 click in the session and 60 seconds of more of time in feed.

Bug: 1377114
Change-Id: I04ddcd14c3c196f502c4455c0a5d35bc166529d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3961654
Commit-Queue: Guillem Perez <guiperez@google.com>
Reviewed-by: Adam Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1067293}
NOKEYCHECK=True
GitOrigin-RevId: b2a9e3981d1b35a12f81c85fda08f193bda9e9e8