[iOS][finch-first-run] Refactor first run and fix bug
In the existing implementation, if `shouldFetchSeed` returns NO and no
variations seed is stored in local_state before the app goes to
background, the next launch would still be considered "first run". To
avoid this, a "placeholder value" is introduced and will be stored in
NSUserDefaults on first run, so the value for "last fetch time" would
not be empty in the next launch.

To make first run utils useful, the code is also refactored so that the
code `ios::RegisterPathProvider()` is invoked on app startup, so the
first run sentinel could be retrieved as soon as possible. This removes
the necessity for the enterprise app agent and the first run app agent
to set `startup_information.isFirstRun`, since it could be set on the
initialization of the main controller.

Bug: 1372180
Change-Id: I11d292626ef383e2b06e304ca5992e211ec89d91
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3998700
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1067563}
NOKEYCHECK=True
GitOrigin-RevId: b88db3e25c7a11c4d745a52bf83f24cdb009d579